Add directive for custom event listeners
When the polymer component updates display count property to a valid
value, the digest cycle runs. But for an invalid value of 0 the
digest cycle does not run and the button is left enabled until the input
field loses focus. With this change, the button is set to disabled as soon
as the input is set to 0.

This is a nice collaboration between Polymer's "notify" feature to send
changed events up to containing elements and angular's custom
directive capability. This custom directive
wouldn't be necessary at all in angular 1.7.3 which includes ngOn.